lantern-dow30/requirements.txt
# Core Airflow and web scraping
apache-airflow==2.10.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
rapidfuzz==3.9.6

# Data processing and utilities
pandas==2.2.2
python-dateutil==2.9.0
pytz==2024.1
tqdm==4.66.5
jsonschema==4.23.0

# Google Cloud Platform integration
google-cloud-storage==2.18.2
google-cloud-documentai==2.30.0
google-api-core==2.19.1

# Database and security
psycopg2-binary==2.9.9
cryptography==41.0.7

# Development and testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Monitoring and logging
prometheus-client==0.19.0